Move api types to common, to avoid ciclic import error
